requirements.txt
pandas
requests
sqlalchemy
pymysql
python-dotenv
pyyaml
matplotlib
seaborn
plotly
pandasgui